+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+### Fixed
+
+- RandomSkunk.Results.Analyzers:
+    - Fix generated method when return type is `ValueTask` or `ValueTask<T>`.
+
 ## [1.0.0-alpha22] - 2022-12-09
 
 ### Changed